Included logic for agent reduction as discussed with Sasha, Deniz and Mit:
  - compute_features.py: Track_id for AV is identified,
interesection of bounding boxes is computed and agents that do not
intersect with the AV are removed.

  - utils/capstone_utils.py: Contains
the helper functions to determine bounding box intersection, and
bounding box for all actors in the scene. Currently, the bounding box
function (bounding_region) is in skeleton form and needs to be
completed.

  - Currently experiencing an error in the code for the func.
get_min_distance_front_and_back on line 248, due to out of bounds.
Unsure if this code section was changed.